Adding output in the Umple documenter to open the selected example code in UmpleOnline. When the user manual is next build, each example will have a link that will automatically load the code into umpleonline.
This is an incremental commit. Before this works really nicely, I will need to add layout information to the examples, so they look good in the layout. Also it turns out that many of the examples are actually bad umple syntax! We really need to arrange to have these built automatically.

git-svn-id: http://umple.googlecode.com/svn/trunk@678 6ebf9d09-8eb1-0036-9d08-68c07d1b990a