This commit contains significant changes to the writeFile() method of the SQLGenerator class. The changes allow for the addition of Foreign key constraints. Thanks to these changes, associations of multiplicity 0..1, 1, and N are now supported. I began to add support for associations of other multiplicities but it is incomplete at the present time.
Also added in this commit is MySQL support for const attributes. MySQL does NOT support the CHECK statement, so additional triggers were required. They have been added.

I have also added the first automated tests for the SQL generation. The tests include association tests for the supported types of associations listed above, as well as a test for the attribute types. The attribute type test should probably be rewritten though, since it does not fully test the SQL generator's attribute generation.

Lastly, as part of making the automated tests work, I had to change every use of the term "SQL" to "Sql" within all the source files. This meant renaming some files as well. The renaming part may be an issue with this commit as I'm not 100% sure if it will work. If not, I'll deal with it immediately. Otherwise, this will probably be my last commit for the official UCOSP Winter 2012 term. After my final exams though, I think I will try to get the other association types working.

git-svn-id: http://umple.googlecode.com/svn/trunk@1687 6ebf9d09-8eb1-0036-9d08-68c07d1b990a