Fixed methods in Abstract class from Interface.
Resolved an issue where Umple was incorrectly generating default implementation of an interface's methods in abstract child classes. In the old behaviour, a class would always be forced to implement the methods from an interface. In the new behaviour, abstract methods will not implement the methods, and instead Umple will search for unimplemented methods for both the current class' parent interfaces and the whole chain of superclasses.

Fixes #608

Signed-off-by: Victoria Lacroix <victoria.a.lacroix@gmail.com>