Static variable issue in the mainMainClass
This issue happens when generating two umple models having different
main functions, because we have a static variable
JavaClassGenerator.mainMainClass, which is not nullified after the
compiler finish generation and start in the next one.

A quick resolve to it to nullify JavaClassGenerator.mainMainClass =
null; before running the next test.