Remove references to absolute system paths in UmpleOnline
Prior to this commit, UmpleOnline used absolute system paths to refer to some of its
runtime dependencies.

This is unnecessarily unportable, since it shouldn't actually matter precisely where these
executables are located as long as they are available on $PATH. It is also inconsistent
since only some executables are referred to in this way.

The prefixes '/usr/bin' and '/usr/local/bin/' are removed from the shell invocations that
have them, allowing UmpleOnline to tolerate variations in the location of the executables
it invokes.