Fixes issue #1084
Implements adding a key to a class in a similar way to the isA relationship. Keys are stored in a map until the rest of the tokens have been parse. They are then added to the class after dependency checks have been made to ensure we do not get caught in an infinite loop.

Adds two tests to verify the changes will fix issue #1084. All existing tests pass without a need to modify.