Split the UmpleOnline Dockerfile to save space
The header of the dockerfile that installs all the dependencies is now
split out into an image called umpleonline-base. It should not need to
change much and saves 200MB from the full UmpleOnline image.

The Travis build has been altered to base its Dockerfile on that image, and
rebuild it on tag builds.