Changes the line number used in Java injections
Using before/after to inject code into methods now uses the line number of the first line of code instead of the line where the before/after statement is declared.
This is done to align it with the strategy imposed by regualr user defined methods.

There were no existing tests for this, so I have modified a couple of existing tests and added 1 new one.

Also fixes a bug where only the first of multiple before/after statements had the line number output to the user.