Fixes bugs in calculating line numbers
Fixes two different small bugs:
For the type A compile time issues, there was no check to see if one of the variables was null before exiting out of a inject statement.

Because an additional line was added to the end of inject statements and a different bug was fixed to change the inject line number comment (#1128), the type B run-time debugging was off by one or two lines.