Added support for highlighting queued states
Previously the state locating function would only look to see if a state's first line began with the state name at position 0. Obviously there are cases where this does not occur, such as queued states. This issue presented itself in the Dishwasher example. To support this, the state locating function will now use a regex search instead of exact string matching with the state name.