Removed debugging messages, note about prev commit
The previous commit was sent without a message, so this will function as its note.

Fixed an error relating to afterEvery guard labels being treated as simply after labels internally. This was done by adding a concession in the search regex that if an "after" is present, is may have "Every" directly after it.