Action.splitStates update + cleanup
Fixed an error where the input variable to a charAt() function was being treated as a string when attempting to modify it via addition, so parseInt() has been included to ensure that an int is passed for addition.

Removed unused variables in Action.highlightTransition() definition.

Replaced currBuffer in Action.selectStateInClass() with the earlier used splitBuffer variable.