First iteration of state diagram menus
This version adds Action.drawInputState(), Action.deleteState(), Action.drawStateMenu(), which function very similarly to their cousins for gv class diagrams. Action.drawStateMenu() is called when a user attempt to call the context menu on a node or cluster, while the page render mode is GvStateDiagram.
Action.drawStateMenu() displays the 4 options for edits, and then calls Action.drawInputState() for edits that require textual input, Action.deleteState() to delete a function, and will a soon to be implemented function to add new transitions.

Furthermore, minor regex changes were made to account for additional edge cases when looking for associations including a target class.