Added SOME plural/singular checking for function names
A few templates will now check if the association specialization has changed the multiplicity, and pluralize the function names accordingly.

Note that there are many other templates that require these checks - this is only the first batch.
However this bug is very rare and other issues are more pressing at this time.