Allowed additional inheritance data in ClassInstanceCountSet. Removed reflexive associations until fixed
ClassInstanceCountSet used to have a map of class names to instance counts, but this was not enough information to resolve inheritance.

The new ClassInstanceCount objects hold the instance count, a flag (if its abstract, interface, etc...), and a set of subclasses.
ClassInstanceCountSet will now hold a map of class names to ClassInstanceCount objects.

Until the reflexive associations can be fixed; they will not be represented in the instance diagram. Technically the generated diagram is still valid, since they always have 0 lower bounds.