Fixed Trait and Random Class bugs
Began ignoring associations to traits. These associations are meant to be copied to classes that implement the trait, but the original copy remains in the model and causes issues.

This has fixed the:
Double association trait bug
Template trait bug

Also fixed an issue with the initial class count randomizer.
Changed the default probabilities for the class counts...