Added an infinite loop check
The code will now check for infinite loops before running the class count algorithm. It will return an error if an infinite loop is detected.

The inheritance resolving code will now only choose "safe" paths. It will only push down instances to subclasses that do not link back to the superclass with >=1 multiplicity (which would potentially result in an infinite loop).