Made the infinite loop detection more lax
The infinite loop detection was too strict on its bounds (>=1), preventing a 1--1 loop between 2 classes.
This is the intended behavior for inheritance resolution (we want more instances to reduce the needed amount of instances); but not intended for regular class behavior (where having a 1--1 loop is allowed)

Now there will be an error for >1 loops with regular classes (instead of >=1).
And >=1 loops will be excluded from inheritance resolution.